backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
langchain==0.3.27
langchain-community==0.3.27
langchain-ollama
chromadb
ollama
unstructured
markdown
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
langchain-experimental
celery
redis
pytest-benchmark
psycopg2-binary
pgvector
sse-starlette
PyPDF2
python-multipart
